Phat refactor
