+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/enchants/enchantments/SoulSystem.kt
@@ -3,13 +3,12 @@ package com.mineinabyss.enchants.enchantments
 import com.mineinabyss.enchants.CustomEnchants
 import com.mineinabyss.enchants.addCustomEnchant
 import com.mineinabyss.enchants.updateEnchantmentLore
-import com.mineinabyss.geary.ecs.accessors.TargetScope
-import com.mineinabyss.geary.ecs.api.annotations.Handler
-import com.mineinabyss.geary.ecs.api.entities.with
-import com.mineinabyss.geary.ecs.api.systems.GearyListener
-import com.mineinabyss.geary.ecs.api.systems.provideDelegate
-import com.mineinabyss.geary.ecs.entities.parent
+import com.mineinabyss.geary.annotations.Handler
+import com.mineinabyss.geary.helpers.parent
+import com.mineinabyss.geary.helpers.with
 import com.mineinabyss.geary.papermc.components.Soulbound
+import com.mineinabyss.geary.systems.GearyListener
+import com.mineinabyss.geary.systems.accessors.TargetScope
 import com.mineinabyss.idofront.entities.toPlayer
 import org.bukkit.entity.Player
 import org.bukkit.inventory.ItemStack
