+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/gondolas/GondolaTracker.kt
@@ -1,10 +1,9 @@
 package com.mineinabyss.gondolas
 
 import com.mineinabyss.components.gondolas.UnlockedGondolas
-import com.mineinabyss.geary.ecs.accessors.TargetScope
-import com.mineinabyss.geary.ecs.api.annotations.Handler
-import com.mineinabyss.geary.ecs.api.systems.GearyListener
-import com.mineinabyss.geary.ecs.api.systems.provideDelegate
+import com.mineinabyss.geary.annotations.Handler
+import com.mineinabyss.geary.systems.GearyListener
+import com.mineinabyss.geary.systems.accessors.TargetScope
 import org.bukkit.entity.Player
 
 class GondolaTracker : GearyListener() {
