+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/gondolas/LoadedGondolas.kt
@@ -1,10 +1,9 @@
 package com.mineinabyss.gondolas
 
 import com.mineinabyss.components.gondolas.Gondola
-import com.mineinabyss.geary.ecs.accessors.TargetScope
-import com.mineinabyss.geary.ecs.api.annotations.Handler
-import com.mineinabyss.geary.ecs.api.systems.GearyListener
-import com.mineinabyss.geary.ecs.api.systems.provideDelegate
+import com.mineinabyss.geary.annotations.Handler
+import com.mineinabyss.geary.systems.GearyListener
+import com.mineinabyss.geary.systems.accessors.TargetScope
 import kotlin.collections.set
 
 object LoadedGondolas : GearyListener() {
