+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/guilds/GuildFeature.kt
@@ -11,7 +11,7 @@ import com.mineinabyss.idofront.plugin.registerEvents
 import com.mineinabyss.mineinabyss.core.AbyssFeature
 import com.mineinabyss.mineinabyss.core.MineInAbyssPlugin
 import com.mineinabyss.mineinabyss.core.commands
-import com.mineinabyss.mineinabyss.extensions.hasGuild
+import com.mineinabyss.guilds.extensions.hasGuild
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import nl.rutgerkok.blocklocker.BlockLockerAPIv2
