+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/guilds/extensions/Guild.kt
@@ -6,8 +6,6 @@ import com.mineinabyss.helpers.MessageQueue
 import com.mineinabyss.idofront.messaging.error
 import com.mineinabyss.idofront.messaging.success
 import com.mineinabyss.mineinabyss.core.AbyssContext
-import com.mineinabyss.mineinabyss.extensions.getGuildName
-import com.mineinabyss.mineinabyss.extensions.getGuildRank
 import org.bukkit.Bukkit
 import org.bukkit.ChatColor
 import org.bukkit.OfflinePlayer
