+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/pins/AbyssalPinBehaviour.kt
@@ -1,13 +1,12 @@
 package com.mineinabyss.pins
 
 import com.mineinabyss.components.pins.PinDrop
-import com.mineinabyss.geary.ecs.accessors.EventScope
-import com.mineinabyss.geary.ecs.accessors.TargetScope
-import com.mineinabyss.geary.ecs.api.annotations.Handler
-import com.mineinabyss.geary.ecs.api.systems.GearyListener
-import com.mineinabyss.geary.ecs.api.systems.provideDelegate
-import com.mineinabyss.geary.ecs.entities.parent
+import com.mineinabyss.geary.annotations.Handler
+import com.mineinabyss.geary.helpers.parent
 import com.mineinabyss.geary.papermc.events.bridge.components.Interacted
+import com.mineinabyss.geary.systems.GearyListener
+import com.mineinabyss.geary.systems.accessors.EventScope
+import com.mineinabyss.geary.systems.accessors.TargetScope
 import com.mineinabyss.looty.ecs.components.itemcontexts.PlayerInventorySlotContext
 
 class AbyssalPinBehaviour : GearyListener() {
