+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/pins/AddPinAction.kt
@@ -2,13 +2,12 @@ package com.mineinabyss.pins
 
 import com.mineinabyss.components.descent.DescentContext
 import com.mineinabyss.components.pins.PinDrop
-import com.mineinabyss.geary.ecs.accessors.EventScope
-import com.mineinabyss.geary.ecs.accessors.TargetScope
-import com.mineinabyss.geary.ecs.accessors.building.get
-import com.mineinabyss.geary.ecs.accessors.building.map
-import com.mineinabyss.geary.ecs.api.annotations.Handler
-import com.mineinabyss.geary.ecs.api.systems.GearyListener
-import com.mineinabyss.geary.ecs.api.systems.provideDelegate
+import com.mineinabyss.geary.annotations.Handler
+import com.mineinabyss.geary.systems.GearyListener
+import com.mineinabyss.geary.systems.accessors.EventScope
+import com.mineinabyss.geary.systems.accessors.TargetScope
+import com.mineinabyss.geary.systems.accessors.building.map
+import com.mineinabyss.geary.systems.accessors.get
 import com.mineinabyss.guiy.inventory.guiy
 import com.mineinabyss.pins.ui.AbyssalPinSelectionMenu
 import kotlinx.serialization.SerialName
