+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/relics/sickle/HarvestAction.kt
@@ -1,15 +1,14 @@
 package com.mineinabyss.relics.sickle
 
 import com.mineinabyss.components.relics.Sickle
-import com.mineinabyss.geary.autoscan.AutoScan
-import com.mineinabyss.geary.ecs.accessors.EventScope
-import com.mineinabyss.geary.ecs.accessors.SourceScope
-import com.mineinabyss.geary.ecs.accessors.TargetScope
-import com.mineinabyss.geary.ecs.accessors.building.get
-import com.mineinabyss.geary.ecs.api.annotations.Handler
-import com.mineinabyss.geary.ecs.api.systems.GearyListener
-import com.mineinabyss.geary.ecs.api.systems.provideDelegate
+import com.mineinabyss.geary.annotations.AutoScan
+import com.mineinabyss.geary.annotations.Handler
 import com.mineinabyss.geary.papermc.events.bridge.components.RightClicked
+import com.mineinabyss.geary.systems.GearyListener
+import com.mineinabyss.geary.systems.accessors.EventScope
+import com.mineinabyss.geary.systems.accessors.SourceScope
+import com.mineinabyss.geary.systems.accessors.TargetScope
+import com.mineinabyss.geary.systems.accessors.get
 import com.mineinabyss.helpers.BlockUtil
 import com.mineinabyss.idofront.items.damage
 import com.mineinabyss.idofront.items.editItemMeta
